requirements.txt
opencv_python==4.5.5.64
Flask==2.1.2
google-cloud-aiplatform==1.13.1
protobuf==3.20.1
gunicorn==20.1.0